+++ profiles/bicycle.lua
@@ -217,7 +217,8 @@ function setup()
     avoid = Set {
       'impassable',
       'construction',
-      'proposed'
+      'proposed',
+      'motorroad'
     }
   }
 end